feat: move intrinsic functions to a separate package
