+++ bundle/default-bundle/src/test/resources/application-local.properties
@@ -9,3 +9,4 @@ camunda.client.auth.token-url=null
 # Access from local c4-connectors storybook
 camunda.endpoints.cors.allowed.origins=http://localhost:6006
 camunda.endpoints.cors.mappings=/inbound-instances/**
+camunda.client.rest-address=http://localhost:8088
