+++ connector-commons/README.md
@@ -1,0 +1,4 @@
+# Connector Commons
+
+Things that are commonly reused across this repository, but probably not intended for use elsewhere
+(e.g. in external connectors).
