+++ connector-commons/intrinsic-functions/src/main/java/io/camunda/connector/intrinsic/DefaultIntrinsicFunctionExecutor.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.intrinsic;
+package io.camunda.connector.intrinsic;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Arrays;
