+++ connector-commons/intrinsic-functions/src/main/java/io/camunda/connector/intrinsic/IntrinsicFunctionParameterBinder.java
@@ -14,14 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.intrinsic;
+package io.camunda.connector.intrinsic;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.annotation.Nullable;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
 import java.util.List;
-import javax.annotation.Nullable;
 
 public class IntrinsicFunctionParameterBinder {
