+++ connector-commons/intrinsic-functions/src/main/java/io/camunda/connector/intrinsic/ServiceLoaderIntrinsicFunctionRegistry.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.intrinsic;
+package io.camunda.connector.intrinsic;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
