+++ connector-commons/intrinsic-functions/src/main/resources/META-INF/services/intrinsic.IntrinsicFunctionProvider
@@ -1,0 +1,4 @@
+io.camunda.connector.intrinsic.functions.Base64Function
+io.camunda.connector.intrinsic.functions.CreateLinkFunction
+io.camunda.connector.intrinsic.functions.GetJsonFunction
+io.camunda.connector.intrinsic.functions.GetTextFunction
