+++ connector-commons/pom.xml
@@ -17,6 +17,7 @@
   <modules>
     <module>connector-object-mapper</module>
     <module>http-client</module>
+    <module>intrinsic-functions</module>
   </modules>
 
 </project>
