--- connector-sdk/document/src/main/resources/META-INF/services/io.camunda.intrinsic.IntrinsicFunctionProvider
@@ -1,4 +1,0 @@
-io.camunda.intrinsic.functions.Base64Function
-io.camunda.intrinsic.functions.CreateLinkFunction
-io.camunda.intrinsic.functions.GetJsonFunction
-io.camunda.intrinsic.functions.GetTextFunction
