+++ connector-sdk/jackson-datatype-document/pom.xml
@@ -22,6 +22,10 @@
       <artifactId>connector-document</artifactId>
     </dependency>
     <dependency>
+      <groupId>io.camunda.connector</groupId>
+      <artifactId>intrinsic-functions</artifactId>
+    </dependency>
+    <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
     </dependency>
