+++ connector-sdk/jackson-datatype-document/src/main/java/io/camunda/connector/document/jackson/deserializer/InputStreamDeserializer.java
@@ -23,7 +23,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import io.camunda.connector.api.document.Document;
 import io.camunda.connector.api.document.DocumentFactory;
 import io.camunda.connector.document.jackson.JacksonModuleDocumentDeserializer.DocumentModuleSettings;
-import io.camunda.intrinsic.IntrinsicFunctionExecutor;
+import io.camunda.connector.intrinsic.IntrinsicFunctionExecutor;
 import java.io.IOException;
 import java.io.InputStream;
