+++ connector-sdk/jackson-datatype-document/src/test/java/io/camunda/connector/document/jackson/DocumentDeserializationTest.java
@@ -27,9 +27,9 @@ import io.camunda.connector.api.document.Document;
 import io.camunda.connector.api.document.DocumentFactory;
 import io.camunda.connector.document.jackson.DocumentReferenceModel.CamundaDocumentMetadataModel;
 import io.camunda.connector.document.jackson.DocumentReferenceModel.CamundaDocumentReferenceModel;
+import io.camunda.connector.intrinsic.IntrinsicFunctionExecutor;
 import io.camunda.document.DocumentFactoryImpl;
 import io.camunda.document.store.CamundaDocumentStore;
-import io.camunda.intrinsic.IntrinsicFunctionExecutor;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
