+++ connector-sdk/jackson-datatype-document/src/test/java/io/camunda/connector/document/jackson/IntrinsicFunctionDeserializationTest.java
@@ -28,10 +28,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import io.camunda.connector.api.document.DocumentFactory;
 import io.camunda.connector.document.jackson.JacksonModuleDocumentDeserializer.DocumentModuleSettings;
+import io.camunda.connector.intrinsic.DefaultIntrinsicFunctionExecutor;
+import io.camunda.connector.intrinsic.IntrinsicFunctionExecutor;
 import io.camunda.document.DocumentFactoryImpl;
 import io.camunda.document.store.CamundaDocumentStore;
-import io.camunda.intrinsic.DefaultIntrinsicFunctionExecutor;
-import io.camunda.intrinsic.IntrinsicFunctionExecutor;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
