+++ connector-sdk/jackson-datatype-document/src/test/resources/META-INF/services/intrinsic.IntrinsicFunctionProvider
